I'm starting to understand that there is a concept of 'enablage' that is essential.
What we were calling the 'protected' set should really be the enabled
set. Then, when deciding what's going to be used on the insertion side
of a reconciliation, we must instantiate terms until only enabled
variables are left. This ensures that when applying the merged patch
we won't trigger any undefined variable errors